Started cleaning up whole_body_controller.hpp.
